Move TeaVM files around
